+++ truffle/CHANGELOG.md
@@ -12,6 +12,7 @@ This changelog summarizes major changes between Truffle versions relevant to lan
 * Removed deprecated methods `TruffleLanguage.parse(Source, Node, String...)` and `TruffleLanguage.evalInContext(Source, Node, MaterializedFrame)` and constructor `RootNode(Class, SourceSection, FrameDescriptor)`.
 * Java Interop now wraps exceptions thrown by Java method invocations in host exceptions.
 * Added [JavaInterop.isHostException](http://graalvm.github.io/graal/truffle/javadoc/com/oracle/truffle/api/interop/java/JavaInterop.html#isHostException-java.lang.Throwable-) and [JavaInterop.asHostException](http://graalvm.github.io/graal/truffle/javadoc/com/oracle/truffle/api/interop/java/JavaInterop.html#asHostException-java.lang.Throwable-) to identify and unwrap host exceptions, respectively.
+* Added support for `TruffleLanguage` context pre-initialization in the native image. To support context pre-initialization a language has to implement the [patchContext](http://graalvm.github.io/graal/truffle/javadoc/com/oracle/truffle/api/TruffleLanguage#patchContext-C-com.oracle.truffle.api.TruffleLanguage.Env-) method.
 * The profiler infrastructure (`CPUSampler`, `CPUTracer` and `MemoryTracer`) moved to a new tools suite.
 * Added [LanguageInfo.isInternal](http://graalvm.github.io/graal/truffle/javadoc/com/oracle/truffle/api/nodes/LanguageInfo.html#isInternal--)
 * Removed special Java interop support for `java.util.Map`.